Initial batch of code for version 1.4.0
This commit greatly cleans up the code in order to make it possible to implement ICE in some reasonable way.

There is currently some test code enabling direct TIC computations that will likely be removed in later revisions, but is here currently to allow ICE to be compared to TIC directly.

This group of commits trades off some small amount of computational performance in curve fitting in exchange for this cleanliness, but it was a necessary trade.